// src/shared/api/model/pagedChatMessagesDTO.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Webchat REST API
 * Webchat application
 * OpenAPI spec version: 3.0
 */
import type { ChatMessageDTO } from './chatMessageDTO';

export interface PagedChatMessagesDTO {
  current_page?: number;
  messages?: ChatMessageDTO[];
  total_pages?: number;
}